renamed sanitize helper → html helper
