+++ src/api/action/profile/get.ts
@@ -10,7 +10,7 @@ import {
     isFollowing,
     isHandle,
 } from '../../../helpers/activitypub/actor';
-import { sanitizeHtml } from '../../../helpers/sanitize';
+import { sanitizeHtml } from '../../../helpers/html';
 import { lookupObject } from '../../../lookup-helpers';
 
 interface Profile {
